AppBarをdefaultに戻す
