+++ front/pages/index.vue
@@ -1,15 +1,12 @@
 <template>
   <div>
-    <AppBar />
     <TaskList :tasks="tasks" />
   </div>
 </template>
 <script>
-import AppBar from '@/components/AppBar'
 import TaskList from '@/components/TaskList'
 export default {
   components: {
-    AppBar,
     TaskList,
   },
   data() {
